backend/blog_project/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via
    #   -r requirements.in
    #   redis
django==4.2.13
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.4.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
redis==5.0.6
    # via
    #   -r requirements.in
    #   django-redis
sqlparse==0.5.0
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   asgiref
